refactor: add index on components folder
