+++ src/components/index.js
@@ -1,0 +1,25 @@
+import AuthRoute from './auth-route';
+import InitializingApp from './initializing-app';
+import LanguageSelector from './language-selector';
+import TopBar from './top-bar';
+import LogoLink from './logo-link';
+import BackButton from './back-button';
+import Message from './message';
+import PrismicContent from './prismic-content';
+import PublicRoute from './public-route';
+import SocialLogin from './social-login';
+import TermsConditionsAndPrivacyPolicy from './terms-conditions-and-privacy-policy';
+
+export {
+    AuthRoute,
+    InitializingApp,
+    LanguageSelector,
+    TopBar,
+    LogoLink,
+    BackButton,
+    Message,
+    PrismicContent,
+    PublicRoute,
+    SocialLogin,
+    TermsConditionsAndPrivacyPolicy,
+};
