+++ src/components/message/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
-import ButtonIcon from 'react-rainbow-components/components/ButtonIcon';
+import { ButtonIcon } from 'react-rainbow-components';
 import CloseIcon from '../icons/close';
 import MessageIcon from './icon';
 import hideMessage from '../../redux/actions/app/hide-message';
