+++ src/components/terms-conditions-and-privacy-policy/index.js
