+++ src/components/top-bar/index.js
@@ -4,17 +4,19 @@ import classnames from 'classnames';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import AvatarMenu from 'react-rainbow-components/components/AvatarMenu';
-import Avatar from 'react-rainbow-components/components/Avatar';
-import MenuDivider from 'react-rainbow-components/components/MenuDivider';
-import MenuItem from 'react-rainbow-components/components/MenuItem';
-import { navigateTo } from '../../../history';
-import { logoutUser } from '../../../redux/actions/authentication';
+import {
+    AvatarMenu,
+    Avatar,
+    MenuDivider,
+    MenuItem,
+} from 'react-rainbow-components';
+import { navigateTo } from '../../history';
+import { logoutUser } from '../../redux/actions/authentication';
 import StyledLanguageSelector from './styled/languageSelector';
-import PowerIcon from '../../../components/icons/power';
-import PencilIcon from '../../../components/icons/pencil';
-import PersonIcon from '../../../components/icons/person';
-import LogoLink from '../../../components/logo-link';
+import PowerIcon from '../icons/power';
+import PencilIcon from '../icons/pencil';
+import PersonIcon from '../icons/person';
+import LogoLink from '../logo-link';
 import './styles.css';
 
 function TopBar(props) {
